Adding dashboard functionality

+++ frontend/src/screens/user.js
@@ -1,5 +1,5 @@
 
-import React, { useState, useEffect, useContext } from 'react'
+import React, { useContext, useEffect, useState } from 'react'
 import {View,Text} from 'react-native'
 
 import { BubbleText, DoneTextBold, ContentJustified, PageTitle, StyledBubbleLarge, StyledButton,StyledButtonText, SubTitle } from '../components/styles';
@@ -7,10 +7,23 @@ import AuthContext from '../utils/auth_context';
 import moment from '../node_modules/moment'
 
 const User = ({route, navigation}) => {
-    const {user,setAuthTokens,setUser,setUserInfo} = useContext(AuthContext)
+    const access = JSON.parse(localStorage.getItem("authTokens"))['access']
+    const {user,setAuthTokens,setUser} = useContext(AuthContext)
+    const [student, setStudent] = useState()
+    const [loading,setloading] = useState(true)
     var date = moment()
     .format('YYYY-MM-DD HH:mm:ss');
     date += '+00:00'
+
+    useEffect(()=>{
+        if (loading){
+            if (user.is_staff){
+            }else{
+            StudentDetails()
+            }
+        }
+
+    },[loading])
 
     const updateUserLastLogin = async ()=>{
         let d = {"last_login" : date}
@@ -18,29 +31,64 @@ const User = ({route, navigation}) => {
         const update = await fetch(url, {
             method: "PUT",
             headers: {
+              "Authorization": `Bearer ${access}`,
               "Content-Type": "application/json"
             },
             body: JSON.stringify(d)
           });
         let r = update.status
         
-    }    
+    }
+    
+    const StudentDetails = async ()=>{
+        const studentUrl = `http://127.0.0.1:8000/students/`+user.user_id +`/`
+        const student_response = await fetch(studentUrl, {
+            method : 'GET',
+            headers :{
+                'Authorization' :`Bearer ${access}`, 
+                'Content-Type' : 'application/json',
+              },
+        })
+        await student_response.json().then(data =>setStudent(data)).catch(error=>setStudent(false))
+        setloading(false)
+        
+    }
+
+    if (!user.is_staff&& !loading){
+        return (
+        <View>
+            <ContentJustified>
+                <PageTitle>User Home</PageTitle>  
+                <SubTitle>{user.first_name} {user.last_name} {`\n`}Level: {student.level}</SubTitle>
+                <SubTitle>Log out?</SubTitle>
+                
+                <StyledButton onPress={()=>{
+                    updateUserLastLogin()
+                    setAuthTokens(null)
+                    setUser(null)
+                    localStorage.clear()
+                    navigation.navigate("Login")
+                    }}><StyledButtonText>Log out</StyledButtonText></StyledButton>
+            </ContentJustified>
+        </View>
+    )}else if (user.is_staff){
         return (
             <View>
                 <ContentJustified>
-                    <PageTitle>User Home</PageTitle>  
+                    <PageTitle>User Home</PageTitle> 
                     <SubTitle>Log out?</SubTitle>
+                    
                     <StyledButton onPress={()=>{
                         updateUserLastLogin()
                         setAuthTokens(null)
                         setUser(null)
-                        setUserInfo(null)
                         localStorage.clear()
                         navigation.navigate("Login")
                         }}><StyledButtonText>Log out</StyledButtonText></StyledButton>
                 </ContentJustified>
             </View>
         )
+    }
     }
 
 export default User;
